requirements.txt
appnope==0.1.3
asgiref==3.7.2
asttokens==2.4.0
attrs==23.1.0
Automat==22.10.0
autopep8==2.0.2
backcall==0.2.0
beautifulsoup4==4.12.2
boto3==1.33.10
botocore==1.33.10
certifi==2023.11.17
cffi==1.15.1
charset-normalizer==3.2.0
constantly==15.1.0
cryptography==41.0.4
cssselect==1.2.0
decorator==5.1.1
distlib==0.3.7
Django==4.2.7
django-cors-headers==4.3.1
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
executing==1.2.0
filelock==3.12.4
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ipython==8.15.0
itemadapter==0.8.0
itemloaders==1.1.0
jedi==0.19.0
jmespath==1.0.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
lxml==4.9.3
matplotlib-inline==0.1.6
mysql-connector-python==8.1.0
packaging==23.1
parsel==1.8.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
Protego==0.3.0
protobuf==4.21.12
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
PyDispatcher==2.0.7
Pygments==2.16.1
PyJWT==2.8.0
pyOpenSSL==23.2.0
pyproject-toml==0.0.10
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
queuelib==1.6.2
referencing==0.30.2
requests==2.31.0
requests-file==1.5.1
rpds-py==0.10.3
s3transfer==0.8.2
scrapeops-scrapy-proxy-sdk==1.0
Scrapy==2.11.0
service-identity==23.1.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.2
tldextract==3.6.0
toml==0.10.2
traitlets==5.10.1
Twisted==22.10.0
typing_extensions==4.8.0
urllib3==2.0.5
virtualenv==20.24.5
w3lib==2.1.2
wcwidth==0.2.6
whitenoise==6.6.0
zope.interface==6.0
